Fix ingest pdf parsing, COOP error and SW fetching
