+++ ROADMAP.md
@@ -65,14 +65,14 @@ All data collection must pass through these legal gates before a user accesses t
 *   [x] **Federation Compliance Matrix:** High-level visual matrix (Green/Amber/Red) tracking SafeSport, background checks, and COPPA 2.0 compliance across every club and coach in the state.
 *   [x] **Tournament Operations & Live Results Hub:** Automated multi-venue bracket scheduling, team registration, digital game sheets, and live scorekeeping that instantly pushes results to the Fan and Player OS.
 
-##### 🏢 EPIC 2: DIRECTOR OS & B2B REVENUE ENGINE
+##### [x] 🏢 EPIC 2: DIRECTOR OS & B2B REVENUE ENGINE
 **Mission:** Multi-sport club scaling, logistical domination, and embedded finance [cite: 755].
 *   **Transaction-Based Monetization:** $0 platform base fee monetized via Stripe Connect transaction fees [cite: 755].
 *   [x] **The Vampire Importer:** Frictionless headless CSV ingestion to steal legacy rosters from competitors [cite: 755].
 *   **Logistics & Field Ops Matrix:** Google Maps API integration with Tomorrow.io Webhooks to auto-lock field status upon lightning strikes [cite: 755].
 *   **Compliance Health Scoring:** Visual indicators tracking VPC and SafeSport statuses at the club level [cite: 755].
 
-##### 📋 EPIC 3: COACH OS (THE SIDELINE SIEM)
+##### [x] 📋 EPIC 3: COACH OS (THE SIDELINE SIEM)
 *   ✅ **Visual Styling Lock (Coach OS)**: Audit Passed.
 **Mission:** Figma-grade tactical tools, RL-driven accountability, and expanded staff controls [cite: 756].
 *   **The Tron War Room:** HTML5 Spatial Drill Designer featuring flawless 1:1 SVG drag-and-drop physics, Vantablack identity discs, and neonBloom light trails using `matrixTransform(getScreenCTM().inverse())` [cite: 756, 1184].
