+++ functions/src/domains/rosterIngestParse.js
@@ -226,7 +226,8 @@ async function parsePdfBase64ToCoachPlayers(contentBase64, apiKey) {
   }
   let pdfText = '';
   try {
-    const pdfData = await pdfParse(pdfBuf);
+    const parseFn = typeof pdfParse === 'function' ? pdfParse : (pdfParse.default || pdfParse.PDFParse);
+    const pdfData = await parseFn(pdfBuf);
     pdfText = pdfData.text;
   } catch (err) {
     throw new Error(`PDF parse failed: ${err.message}`);
