--- test-pdf-parse.cjs
@@ -1,2 +1,0 @@
-const pdfParse = require('pdf-parse');
-console.log('pdfParse loaded:', typeof pdfParse);
